# apiserver/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a apiserver for KubeAGI.",
        "title": "ApiServer Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/chat": {
            "post": {
                "description": "chat with application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "chat with application",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Should the chat request be treated as debugging?",
                        "name": "debug",
                        "in": "query"
                    },
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat.ChatReqBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "blocking mode, will return all field; streaming mode, only conversation_id, message and created_at will be returned",
                        "schema": {
                            "$ref": "#/definitions/chat.ChatRespBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    }
                }
            }
        },
        "/chat/conversations": {
            "post": {
                "description": "list all conversations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "list all conversations",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat.APPMetadata"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.Conversation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    }
                }
            }
        },
        "/chat/conversations/:conversationID": {
            "delete": {
                "description": "delete one conversation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "delete one conversation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "conversationID",
                        "name": "conversationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chat.SimpleResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    }
                }
            }
        },
        "/chat/conversations/docs": {
            "post": {
                "description": "receive and process uploaded documents(pdf, docx) for one conversation",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "receive and process uploaded documents(pdf, docx) for one conversation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The app namespace for this conversation",
                        "name": "app_namespace",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The app name for this conversation",
                        "name": "app_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The conversation id for this document",
                        "name": "conversation_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The chunk size when load and split the document",
                        "name": "chunk_size",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "The chunk overlap when load and split the document",
                        "name": "chunk_overlap",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "This is the docs for the conversation",
                        "name": "docs",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chat.ConversationDocsRespBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    }
                }
            }
        },
        "/chat/messages": {
            "post": {
                "description": "get all messages history for one conversation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "get all messages history for one conversation",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat.ConversationReqBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Conversation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    }
                }
            }
        },
        "/chat/messages/:messageID/references": {
            "post": {
                "description": "get one message's references",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "get one message references",
                "parameters": [
                    {
                        "type": "string",
                        "description": "messageID",
                        "name": "messageID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat.MessageReqBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/retriever.Reference"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    }
                }
            }
        },
        "/chat/prompt-starter": {
            "post": {
                "description": "get app's prompt starters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "get app's prompt starters",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "how many prompts you need should \u003e 0 and \u003c 10",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat.APPMetadata"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/chat.ErrorResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "chat.APPMetadata": {
            "type": "object",
            "required": [
                "app_name",
                "app_namespace"
            ],
            "properties": {
                "app_name": {
                    "description": "AppName, the name of the application",
                    "type": "string",
                    "example": "chat-with-llm"
                },
                "app_namespace": {
                    "description": "AppNamespace, the namespace of the application",
                    "type": "string",
                    "example": "kubeagi-system"
                }
            }
        },
        "chat.ChatReqBody": {
            "type": "object",
            "required": [
                "app_name",
                "app_namespace",
                "query",
                "response_mode"
            ],
            "properties": {
                "app_name": {
                    "description": "AppName, the name of the application",
                    "type": "string",
                    "example": "chat-with-llm"
                },
                "app_namespace": {
                    "description": "AppNamespace, the namespace of the application",
                    "type": "string",
                    "example": "kubeagi-system"
                },
                "conversation_id": {
                    "description": "ConversationID, if it is empty, a new conversation will be created",
                    "type": "string",
                    "example": "5a41f3ca-763b-41ec-91c3-4bbbb00736d0"
                },
                "query": {
                    "description": "Query user query string",
                    "type": "string",
                    "example": "旷工最小计算单位为多少天？"
                },
                "response_mode": {
                    "description": "ResponseMode:\n* Blocking - means the response is returned in a blocking manner\n* Streaming - means the response will use Server-Sent Events",
                    "allOf": [
                        {
                            "$ref": "#/definitions/chat.ResponseMode"
                        }
                    ],
                    "example": "blocking"
                }
            }
        },
        "chat.ChatRespBody": {
            "type": "object",
            "properties": {
                "conversation_id": {
                    "type": "string",
                    "example": "5a41f3ca-763b-41ec-91c3-4bbbb00736d0"
                },
                "created_at": {
                    "description": "CreatedAt is the time when the message is created",
                    "type": "string",
                    "example": "2023-12-21T10:21:06.389359092+08:00"
                },
                "latency": {
                    "description": "Latency(ms) is how much time the server cost to process a certain request.",
                    "type": "integer",
                    "example": 1000
                },
                "message": {
                    "description": "Message is what AI say",
                    "type": "string",
                    "example": "旷工最小计算单位为0.5天。"
                },
                "message_id": {
                    "type": "string",
                    "example": "4f3546dd-5404-4bf8-a3bc-4fa3f9a7ba24"
                },
                "references": {
                    "description": "References is the list of references",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/retriever.Reference"
                    }
                }
            }
        },
...
        "chat.ConversationDocsRespBody": {
            "type": "object",
            "properties": {
                "docs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/chat.ConversatioSingleDocRespBody"
                    }
                }
            }
        },
...
        "chat.ErrorResp": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "conversation is not found"
                }
            }
        },
...
        "chat.SimpleResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "storage.Conversation": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string",
                    "example": "chat-with-llm"
                },
                "app_namespace": {
                    "type": "string",
                    "example": "arcadia"
                },
                "id": {
                    "type": "string",
                    "example": "5a41f3ca-763b-41ec-91c3-4bbbb00736d0"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.Message"
                    }
                },
                "started_at": {
                    "type": "string",
                    "example": "2023-12-21T10:21:06.389359092+08:00"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-12-22T10:21:06.389359092+08:00"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API token for authorization",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}